TASK-1107 Use configuration passed on init
Upstream, the SparkpostRails gem allows configuration one of two ways
 - via an initializer that calls `SparkPostRails.configure { |c| ... }`
 - via individual mailer methods, e.g. `MyMailer.password_reset {
 sparkpost_data: { ... }}`

The latter overrides anything set in the former.

The ActionMailer interface provides a third method that is only
partially supported in the upstream version of this gem. Each
`Mail::Message` instance contains a dup of the `Mail.delivery_method`
settings ([1][1]). That allows for the delivery method to be modified
after the message is generated but before it is delivered, e.g.

    message = MyMailer.password_reset
    message.delivery_method SparkPostRails::DeliveryMethod, { ... }

SparkPostRails::Delivery method allows for the settings to be passed in,
but never uses them ([3][3]).

This changeset adds a very simple fallback in the case that _no_
settings are found in the mailer call[4][4], the DeliveryMethod will use
any settings passed when the instance was created.

We need this only temporarily. Once our migration to Sparkpost is
complete, we can drop our fork, use the upstream gem, and rely on the
app-wide configuration with mailer method overrides for categories and
things of that sort.

[1]: https://github.com/mikel/mail/blob/aba0b5fd3a073fea1806eb3ac2aaf640941f7362/lib/mail/message.rb#L126
[2]: https://github.com/mikel/mail/blob/aba0b5fd3a073fea1806eb3ac2aaf640941f7362/lib/mail/message.rb#L278
[3]: https://github.com/the-refinery/sparkpost_rails/blob/1de062ec0c6a8608d9d011b18239bfeb731ab84a/lib/sparkpost_rails/delivery_method.rb#L8
[4]: https://github.com/the-refinery/sparkpost_rails/blob/1de062ec0c6a8608d9d011b18239bfeb731ab84a/lib/sparkpost_rails/delivery_method.rb#L14